# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dispatch/suggestions/{jobId}:
    get:
      operationId: getDispatchSuggestions
      tags: [dispatch]
      summary: Get ranked subcontractor suggestions for a job
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranked suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DispatchSuggestion"
  /client/lookup:
    post:
      operationId: clientLookupJob
      tags: [client]
      summary: Look up a job by phone number and job ID for the client portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientLookupBody"
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
  /client/request:
    post:
      operationId: clientRequestService
      tags: [client]
      summary: Submit a new service request from the client portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Request received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /escalations:
    get:
      operationId: listEscalations
      tags: [escalations]
      summary: List all escalations
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of escalations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escalation"
    post:
      operationId: createEscalation
      tags: [escalations]
      summary: Create an escalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscalationBody"
      responses:
        "201":
          description: Created escalation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
  /escalations/{id}:
    patch:
      operationId: updateEscalation
      tags: [escalations]
      summary: Update escalation status or notes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEscalationBody"
      responses:
        "200":
          description: Updated escalation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
        "404":
          description: Not found
...
  /invoices/job/{jobId}:
    get:
      operationId: getInvoiceByJob
      tags: [invoices]
      summary: Get invoice for a job
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceStatusBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices:
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
components:
  schemas:
    ClientLookupBody:
      type: object
      properties:
        phone:
          type: string
        jobId:
          type: integer
      required:
        - phone
        - jobId
...
    Job:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
        planName:
          type: ["string", "null"]
        customerPrice:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        assignment:
          $ref: "#/components/schemas/Assignment"
          nullable: true
      required:
        - id
        - customerName
        - phone
        - address
        - serviceType
        - status
        - createdAt
    CreateJobBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
        planName:
          type: ["string", "null"]
        customerPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - phone
        - address
        - serviceType
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        subcontractorId:
          type: integer
        subName:
          type: ["string", "null"]
        subPay:
          type: number
        ownerProfit:
          type: number
        status:
          type: string
        assignedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - subcontractorId
        - subPay
        - ownerProfit
        - status
        - assignedAt
...
    DispatchSuggestion:
      type: object
      properties:
        subcontractorId:
          type: integer
        name:
          type: string
        phone:
          type: string
        specialty:
          type: string
        activeJobs:
          type: integer
        jobsToday:
          type: integer
        isOverbooked:
          type: boolean
        specialtyMatch:
          type: boolean
        score:
          type: integer
        reason:
          type: string
      required:
        - subcontractorId
        - name
        - phone
        - specialty
        - activeJobs
        - jobsToday
        - isOverbooked
        - specialtyMatch
        - score
        - reason
...
    Escalation:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        category:
          type: string
        description:
          type: string
        status:
          type: string
        resolution:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - category
        - description
        - status
        - createdAt
        - updatedAt
    CreateEscalationBody:
      type: object
      properties:
        jobId:
          type: ["integer", "null"]
        category:
          type: string
        description:
          type: string
      required:
        - category
        - description
    UpdateEscalationBody:
      type: object
      properties:
        status:
          type: string
        resolution:
          type: ["string", "null"]
...
    InvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - amount
    Invoice:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        invoiceNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        serviceType:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        sentAt:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - invoiceNumber
        - customerName
        - customerPhone
        - serviceType
        - lineItems
        - subtotal
        - tax
        - total
        - status
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        jobId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        serviceType:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - jobId
        - customerName
        - customerPhone
        - serviceType
        - lineItems
        - subtotal
        - total
    UpdateInvoiceStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status